Fix post deletion. Fixes #8

+++ app/lib/micropub/delete.js
@@ -12,34 +12,30 @@ const response = require(__basedir + '/lib/micropub/response');
  * @returns {Promise} Response object
  */
 module.exports = async url => {
-  let repoPath;
-  let entries;
+  let historyEntry;
 
   try {
     const getHistory = await history.read();
-    entries = getHistory.entries;
+    const {entries} = getHistory;
+    historyEntry = entries.filter(entry => entry.create.url === url);
   } catch (error) {
     console.error(error);
   }
 
-  if (entries) {
-    entries.forEach(entry => {
-      if (entry.create.url === url) {
-        repoPath = entry.create.post;
+  if (historyEntry) {
+    try {
+      const repoPath = historyEntry[0].create.post;
+      const githubResponse = await github.deleteFile(repoPath, {
+        message: `:x: Post deleted with ${appConfig.name}`
+      });
+      if (githubResponse) {
+        // @todo Save properties to history to enable undelete action
+        return response.success('delete', url);
       }
-
-      console.info(`${url} not found in history`);
-    });
-  }
-
-  try {
-    const githubResponse = await github.deleteFile(repoPath, {
-      message: `:x: Post deleted with ${appConfig.name}`
-    });
-    if (githubResponse) {
-      return response.success('delete', url);
+    } catch (error) {
+      throw new Error(`Unable to delete ${url}. ${error.message}`);
     }
-  } catch (error) {
-    throw new Error(`Unable to delete ${url}`);
   }
+
+  return response.error('not_found');
 };
